refactor(custom_formats): split 2319-line file into mod/parser/evaluator
Mirrors the services/source/ precedent: mod.rs aggregates and re-exports
the public surface, siblings own one concern each. Parser (compile_from_json,
sonarr_to_rust_regex, field_* helpers) and evaluator (evaluate_spec_kernel,
evaluate, total_cf_score, total_cf_score_with_breakdown) are genuinely
decoupled — the scoring hot path in auto_search.rs only touches evaluator
output once parsing completed at startup, so the split fractures along
the natural boundary. Shared test helpers (candidate/classification/ctx/
compile fixtures) live in mod.rs::test_helpers so both siblings' test
modules can share them without duplication.

No behavior change. External call sites (auto_search.rs, main.rs,
upgrade.rs, seadex.rs) all continue to resolve because everything public
is either still in mod.rs (types, DB loader, has_seadex_cf) or re-exported
via `pub use` (compile_from_json, total_cf_score_with_breakdown). The
`evaluate` and `total_cf_score` re-exports were dropped because nothing
in-crate references them at the custom_formats::X path — they remain
`pub fn` in evaluator.rs for future callers that want them back.
